#280 Car turning fix

+++ src/main/java/hu/oe/nik/szfmv/Main.java
@@ -50,7 +50,7 @@ public class Main {
                 pedestrian.moveOnCrosswalk();
 
                 gui.getCourseDisplay().drawWorld(w, car.getCarValues());
-                gui.getDashboard().updateDisplayedValues(car.getInputValues(), car.getX(), car.getY());
+                gui.getDashboard().updateDisplayedValues(car.getInputValues(), (int) car.getX(), (int) car.getY());
                 Thread.sleep(CYCLE_PERIOD);
             } catch (InterruptedException e) {
                 LOGGER.error(e.getMessage());
